+++ packages/lib/src/propTransform/transformArrayAsMap.ts
@@ -1,28 +1,126 @@
+import {
+  action,
+  IArrayChange,
+  IArraySplice,
+  IMapWillChange,
+  intercept,
+  IObservableArray,
+  isObservableArray,
+  observable,
+  ObservableMap,
+  observe,
+} from "mobx"
 import { MaybeOptionalModelProp, OnlyPrimitives, OptionalModelProp, prop } from "../model/prop"
 import { AnyType, TypeToData } from "../typeChecking/schemas"
 import { tProp } from "../typeChecking/tProp"
-import { isArray, isMap } from "../utils"
-import { arrayAsMap } from "../wrappers"
+import { failure, inDevMode, isArray, isMap } from "../utils"
+import { Lock } from "../utils/lock"
 import { PropTransform, transformedProp } from "./propTransform"
 
-/**
- * A map expressed as an array.
- */
-export type TransformArrayAsMap<V> = [string, V][]
+const observableMapBackedByObservableArray = action(
+  <T>(array: IObservableArray<[string, T]>): ObservableMap<string, T> => {
+    if (inDevMode()) {
+      if (!isObservableArray(array)) {
+        throw failure("assertion failed: expected an observable array")
+      }
+    }
+
+    const map = observable.map(array)
+
+    if (map.size !== array.length) {
+      throw failure("arrays backing a map cannot contain duplicate keys")
+    }
+
+    const mutationLock = new Lock()
+
+    // for speed reasons we will just assume distinct values are only once in the array
+    // also we assume tuples themselves are immutable
+
+    // when the array changes the map changes
+    observe(
+      array,
+      action(
+        mutationLock.unlockedFn((change: IArrayChange<[string, T]> | IArraySplice<[string, T]>) => {
+          switch (change.type) {
+            case "splice": {
+              {
+                const removed = change.removed
+                for (let i = 0; i < removed.length; i++) {
+                  map.delete(removed[i][0])
+                }
+              }
+
+              {
+                const added = change.added
+                for (let i = 0; i < added.length; i++) {
+                  map.set(added[i][0], added[i][1])
+                }
+              }
+
+              break
+            }
+
+            case "update": {
+              map.delete(change.oldValue[0])
+              map.set(change.newValue[0], change.newValue[1])
+              break
+            }
+          }
+        })
+      )
+    )
+
+    // when the map changes also change the array
+    intercept(
+      map,
+      action((change: IMapWillChange<string, T>) => {
+        if (!mutationLock.isLocked) {
+          return null // already changed
+        }
+
+        switch (change.type) {
+          case "update": {
+            // replace the whole tuple to keep tuple immutability
+            const i = array.findIndex(i => i[0] === change.name)
+            array[i] = [change.name, change.newValue!]
+            break
+          }
+
+          case "add": {
+            array.push([change.name, change.newValue!])
+            break
+          }
+
+          case "delete": {
+            const i = array.findIndex(i => i[0] === change.name)
+            if (i >= 0) {
+              array.splice(i, 1)
+            }
+            break
+          }
+        }
+
+        return change
+      })
+    )
+
+    return map
+  }
+)
 
 const arrayAsMapInnerTransform: PropTransform<
   [string, any][] | unknown,
   Map<string, any> | unknown
 > = {
   propToData(arr) {
-    return isArray(arr) ? arrayAsMap(() => arr) : arr
+    return isArray(arr) ? observableMapBackedByObservableArray(arr as IObservableArray) : arr
   },
   dataToProp(newMap) {
     if (!isMap(newMap)) {
       return newMap
     }
 
-    const arr: TransformArrayAsMap<any> = []
+    const arr: [string, any][] = []
     for (const k of newMap.keys()) {
       arr.push([k, newMap.get(k)])
     }