+++ packages/lib/test/propTransform/implicit/transformTimestampAsDate.test.ts
@@ -32,6 +32,11 @@ test("transformTimestampAsDate", () => {
     setTimestamp(timestamp: number) {
       this.$.date = timestamp
     }
+
+    @modelAction
+    setDateTime(time: number) {
+      this.date.setTime(time)
+    }
   }
 
   assert(_ as SnapshotInOf<M>["date"], _ as number)
@@ -46,7 +51,7 @@ test("transformTimestampAsDate", () => {
 
   // getter
   expect(m.date instanceof Date).toBeTruthy()
-  expect(m.date).toBe(dateNow) // must be the same instance
+  expect(m.date).toEqual(dateNow) // not same instance, transformation will generate a new one
 
   // should be cached
   expect(m.date).toBe(m.date)
@@ -84,7 +89,7 @@ test("transformTimestampAsDate", () => {
   expect(timestampAsDate.dataToProp(dateNow2)).toStrictEqual(now2)
 
   m.setDate(dateNow2)
-  expect(m.date).toBe(dateNow2)
+  expect(m.date).toEqual(dateNow2)
   expect(m.$.date).toBe(now2)
 
   expect(actionCalls).toMatchInlineSnapshot(`
@@ -128,13 +133,13 @@ test("transformTimestampAsDate", () => {
   // changing the date to be the same should keep the cached value intact
   reactions.length = 0
   m.setDate(dateNow)
-  expect(m.date).toBe(dateNow)
+  expect(m.date).toEqual(dateNow)
   expect(reactions).toHaveLength(0)
 
   // changing the backed value to be the same should keep the cached value intact
   reactions.length = 0
   m.setTimestamp(now)
-  expect(m.date).toBe(dateNow)
+  expect(m.date).toEqual(dateNow)
   expect(reactions).toHaveLength(0)
 
   // changing the backed prop and trying to set the same data back should work
@@ -152,4 +157,8 @@ test("transformTimestampAsDate", () => {
       1970-01-01T00:00:10.000Z,
     ]
   `)
+
+  // make sure mutations on the date object are picked up in the backed prop
+  m.setDateTime(10)
+  expect(m.$.date).toBe(10)
 })